refactor(search-help-posts): add new package search
